Add feat delete account user

+++ routes/api.php
@@ -4,25 +4,11 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\UserController;
 
-// Route untuk menghapus user secara permanen
-Route::delete('/users/{id}/permanently', [UserController::class, 'destroyPermanently']);
-
-
+// Rute untuk menghapus user secara permanen dengan middleware autentikasi
+Route::middleware('auth:sanctum')->delete('/users/{id}/permanently', [UserController::class, 'destroyPermanently']);
 Route::get('/users', [UserController::class, 'index']);
 
-
-
-/*
-|--------------------------------------------------------------------------
-| API Routes
-|--------------------------------------------------------------------------
-|
-| Here is where you can register API routes for your application. These
-| routes are loaded by the RouteServiceProvider within a group which
-| is assigned the "api" middleware group. Enjoy building your API!
-|
-*/
-
+// Rute untuk mendapatkan data user yang terautentikasi
 Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
